finish part4

+++ part4/bloglist/controllers/blogs.js
@@ -1,17 +1,17 @@
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blogs');
+const User = require('../models/users');
+const jwt = require('jsonwebtoken');
 
-
-blogsRouter.get('/', (request, response) => {
-  Blog
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog
     .find({})
-    .then(blogs => {
-      response.json(blogs);
-    });
+    .populate('user', { username: 1, name: 1 });
+  response.json(blogs);
 });
 
-blogsRouter.get('/:id', (request, response) => {
-  const blog = Blog.findById(request.params.id);
+blogsRouter.get('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id);
   if (blog) {
     response.json(blog);
   } else {
@@ -20,8 +20,9 @@ blogsRouter.get('/:id', (request, response) => {
 });
 
 // Make Undefined likes to be 0
-blogsRouter.post('/', (request, response, next) => {
+blogsRouter.post('/', async (request, response) => {
   const body = request.body;
+  const user = request.user;
 
   if (body.title == undefined || body.url == undefined) {
     return response.status(400).end();
@@ -31,14 +32,13 @@ blogsRouter.post('/', (request, response, next) => {
       author: body.author,
       url: body.url,
       likes: body?.likes | 0,
+      user: user.id
     });
-  
-    blog
-      .save()
-      .then(result => {
-        response.status(201).json(result);
-      })
-      .catch(error => next(error));
+
+    const savedblog = await blog.save();
+    user.blogs = user.blogs.concat(savedblog._id);
+    await user.save();
+    response.status(201).json(savedblog);
   }
 });
 
@@ -59,12 +59,17 @@ blogsRouter.put('/:id', async (request, response) => {
 
 
 blogsRouter.delete('/:id', async (request, response) => {
-  const deletedBlog = await Blog.findByIdAndDelete(request.params.id);
-  if (!deletedBlog) {
-    // No blog found with the given ID
-    return response.status(404).send({ error: 'Blog not found' });
+  const user = request.user; 
+  if (!user){
+    return response.status(401).json({ error: 'Unauthorized to delete the blog' }); 
   }
-  response.status(204).end();
+  const blog = await Blog.findById(request.params.id);
+  if (blog.user.toString() == user.id) {
+    await Blog.findByIdAndDelete(request.params.id);
+    return response.status(204).send({ error: 'Blog not found' });
+  } else {
+    return response.status(401).json({ error: 'Unauthorized to delete the blog' });
+  }
 });
 
 module.exports = blogsRouter;
